requirements.txt
beautifulsoup4==4.12.3
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
googletrans==4.0.0rc1
gTTS==2.5.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.4.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
joblib==1.4.0
nltk==3.8.1
pyTelegramBotAPI==4.16.1
regex==2023.12.25
requests==2.31.0
rfc3986==1.5.0
sniffio==1.3.1
soupsieve==2.5
tqdm==4.66.2
urllib3==2.2.1
eng_to_ipa==0.0.2
mysql-connector-python==8.3.0
matplotlib==3.5.3
pytz

